MIPS: lantiq: Allow specifying the reset controller register offsets
This changes of_reset_n_cells to 2 since the reset controller has
multiple registers of which each bit is assigned to different functionality
(USB, GPHY, etc.). This allows us to specify the register and it's bits via
devicetree.
Adding backwards compatibility (for of_reset_n_cells = 1) is not required
since the reset_controller_dev was not registered before (ltq_rst_init
was uncalled).