mmc: meson-gxbb: experimental SDIO support
This adds experimental SDIO support to the meson mmc controller.
The patch consists of four main changes:
- remove meson_mmc_check_cmd to allow the SDIO commands
- implement meson_mcc_enable_sdio_irq - this requires adding the
  irq_mask field to our priv struct so we always configure the
  correct IRQs.
- get rid of the magic value IRQ_EN_MASK and replace it with an OR
  of the corresponding bit/masks (when initializing host->irq_mask)
- it adds a small bugfix in meson_mmc_probe when probing failed (in
  my case: probing had to be deferred because the mmc-initseq was not
  ready yet) which disabled and unprepared a clock which was not
  enabled

I am sending this patch because I only have limited access to my board
in the next 2 weeks. Please see this as "proof of concept" and not a
finished patch. If someone wants to work in this: feel free - otherwise
I'll do it when I have full access to my board again.

An additional patch which enables wifi on the Tronsmart S95 Vega Meta
can be found here:
https://github.com/xdarklight/linux/tree/meson-gxbb-integration-4.8-20160828

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>