mmc: meson-gxbb: add support for the SDIO interrupt
The first MMC host controller (sd_emmc_a) has support for SDIO
interrupts (on a dedicated pin). This is signalled by the IRQ_SDIO bit
in the SD_EMMC_STATUS register.
While here this also removes the IRQ_EN_MASK definition and replaces it
with an easier to read variant (which lists all bits which are enabled
using meaningful names).

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>