phy: meson: add USB2 and USB3 PHY support for Meson GXL and GXM
This adds two new USB PHY drivers found on Meson GXL and GXM SoCs (as
both SoCs are using the same USB PHYs).

The USB2 PHY is a simple PHY which only has a few registers to configure
the mode (host, OTG, device) and a reset register (to enable/disable the
PHY).

The USB3 PHY actually has three purposes:
- it provides the USB3 PHY
- it handles the OTG device/host mode detection interrupt
- it notifies the corresponding USB2 PHYs of the OTG mode changes
On GXL and GXM SoCs one references all available USB2 PHY ports in the
USB3 PHY because all are connected to the same USB controller (thus the
mode will always match). This behavior is configurable via devicetree,
by passing (or not passing) a list of other ("child") PHYs which should
be configured by the USB3 PHY.

Unfortunately there are no datasheets available for any of these PHYs.
Both drivers were written by reading the reference drivers provided by
Amlogic and analyzing the registers on the kernel that was shipped on
the boards I have.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>