usb: dwc3: make it possible to configure multiple USB PHYs
The DWC3 hardware has an internal USB hub. On most devices only one port
is enabled, for these the current PHY configuration (defining the PHYs
directly in the dwc3 devicetree node) is perfectly fine. However, there
are some devices (for example the Amlogic GXL and GXM SoCs) which have
more than one port on the internal hub enabled. Those devices require
specifying one PHY per port on the root-hub.
Use the new generic PHY hub driver to enable all PHYs on the internal
hub when the root-hub and it's ports are defined in the devicetree.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>