phy: meson8b-usb2: unset the IDDQ bit in host mode during PHY power-on
The vendor driver does this as well and unsets it when suspending the
PHY. I did not notice this in the vendor driver first, because it's part
of the dwc_otg driver there (instead of their PHY code).
While here, also add all other REG_DBG_UART register bit definitions.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>