iio: adc: meson-saradc: add calibration for meson gxbb
The calibration on Meson GXBB is different from the one on GXM/GXL.
This patch only supports GXBB as I don't have other spec and test hw.
Points 25% vref and 75% vref are used for calibration.

Adding a calibration scale resulted in no improvement, therefore
only a simple calibration bias is supported.

Successfully tested on a Odroid C2.  Calibbias is 8 on my system.

Signed-off-by: Heiner Kallweit <hkallweit1@gmail.com>