net: stmmac: dwmac-meson8b: fix setting the PHY clock on Meson8b
Meson8b uses fclk_div5 as one of the clock inputs. The rate of that
clock is 1275MHz. Setting the PHY clock rate fails on Meson8b because
the wanted clock rate (25MHz for RGMII PHYs and clk_get_rate(fclk_div5)
for RMII PHYs) cannot be set without rounding the clock rate to the
closest possible value.
This has no effect on the Meson GX SoCs since there fclk_div2 is used as
input clock, which has a rate of 1000MHz (and thus is divisible cleanly
to 25MHz).

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>