clk: meson: meson8b: add the reset controller inside the clock registers
This adds the reset controller for the CPU cores (which are needed to
turn individual cores on or off, by asserting or de-asserting the
corresponding reset line).

According to the datasheet there are more reset lines inside
HHI_SYS_CPU_CLK_CNTL0:
- bit 14 = SOFT_RESET
- bit 15 = GEN_DIV_SOFT_RESET
- bit 16 = A5 APB Soft Reset
- bit 17 = A5 AXI Soft Reset
- bit 18 = A5 Global Reset
- bit 28 = SCU soft reset
- bit 29 = AXI64to128 bridge (A5-to-MMC) soft reset (A5 interface)
- bit 30 = L2 cache soft reset
However, these are not used in the Amlogic GPL kernel sources. Instead
they are only used by u-boot and/or the suspend firmware. If required,
these can be added in the future.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>