clk: meson: meson8b: add compatibles for Meson8 and Meson8m2
The clock controller on Meson8, Meson8b and Meson8m2 seems identical
based on the Amlogic GPL kernel sources. Add separate compatibles for
each SoC to make sure that we can easily add small changes for each
SoC if we ever find out that I was wrong and there are differences.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>