ARM: dts: meson: add the MMC controller
Meson6, Meson8 and Meson8b are using the same MMC controller IP. This
adds the MMC controller node to meson.dtsi so it can be used by all
SoCs.

The controller itself is a bit special, because it has multiple slots.
Each slot is accessed through a sub-node of the controller. To simplify
the device .dts files each SoC enables the MMC controller node directly
so the devices only have to enable the corresponding slots.
On Meson6 the controller is not enabled because I don't have hardware to
test it there.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>