pwm: meson: add support for the Meson8 SoC PWM controller
The PWM controllers in the Meson8 SoCs use the same register layout as
already implemented by the pwm-meson driver. The clock controller in the
Meson8 SoCs is the same as the one found in Meson8b SoCs. Add a new
compatible to the driver and re-use (and rename) the existing parent
clocks to make the pwm-meson driver compatible with Meson8 SoCs.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>