pwm: meson: add Meson8 and Meson8m2 support
The PWM controllers in the Meson8 and Meson8m2 SoCs use the same register
layout as already implemented by the pwm-meson driver. As the clock
controller in these SoCs is the same as the one found in Meson8b SoCs, add
new of_device_ids to the driver and re-use (and rename) the existing parent
clocks.
This makes the pwm-meson driver compatible with Meson8 and Meson8m2 SoCs.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>