iio: adc: meson-saradc: add support for the chip's temperature sensor
Channel 6 of the SAR ADC can be switched between two inputs:
SAR_ADC_CH6 input (an actual pad on the SoC) and the signal from the
temperature sensor inside the SoC.

To get usable results from the temperature sensor we need to read the
corresponding calibration data from the eFuse and pass it to the SAR ADC
(bits [3:0]) and (on Meson8b and Meson8m2) to a scratch register in the
HHI region.
If the temperature sensor is not calibrated (the eFuse data contains a
bit for this) then the driver will return -ENOTSUPP when trying to read
the temperature.

This only enables the temperature sensor for the Meson8, Meson8b and
Meson8m2 SoCs because on the newer, 64-bit SoCs (GXBB, GXL and GXM) the
temperature sensor inside SAR ADC is firmware-controlled (by BL30, we
can simply use the SCPI hwmon driver to get the chip temperature).

To keep the devicetree interface backwards compatible we simply skip the
temperature sensor initialization if no eFuse nvmem cell is passed via
devicetree.

The public documentation for the SAR ADC IP block does not explain how
to use the registers to read the temperature. The logic from this patch
is based on reading and understanding Amlogic's GPL kernel sources.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>