iio: adc: meson-saradc: add support for the temperature sensor - WiP
On the meson-saradc channel 6 can be switched between two inputs: the
actual SAR_ADC_CH6 input (an actual pad on the SoC) and the signal from
the temperature sensor inside the SoC.

To get usable results from the temperature sensor we need to read the
corresponding calibration data from the efuse.
Not all boards come with the required calibration data. The temperature
sensor channel is disabled automatically when the calibaration data is
missing.

Right now we only enable the temperature sensor on Meson8 and Meson8b
SoCs since the logic for that is pretty simple and it does not involve
any registers outside the SAR ADC. For GXBB and newer the logic is much
more complex and seems to involve some registers that are not documented
in public datasheets. We can also simply use the PSCI hwmon driver on
GXBB and newer to read the temperature sensor instead (in this case the
BL30 firmware uses the SAR ADC to read the SoC temperature).

To keep the devicetree interface backwards compatible we simply skip the
temperature sensor initialization if no efuse nvmem cell is passed via
devicetree.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>