iio: adc: meson-saradc: improve support for older SoCs
There are some small differences between the old (32-bit) Meson8 and
Meson8b SoCs and the newer (64-bit) Meson GX (GXBB, GXL, GXM) SoCs:
- older SoCs do not have MESON_SAR_ADC_REG11 and MESON_SAR_ADC_REG13
  thus we limit the regmap so MESON_SAR_ADC_DELTA_10 is our max register
- GX SoCs use the bandgap setting from MESON_SAR_ADC_REG11, while the
  older SoCs use the TS_VBG_EN bit in MESON_SAR_ADC_DELTA_10
- GX SoCs use a 1.2 MHz ADC clock, while the older SoCs use a 1.14 MHz
  clock (although the comment in the driver from Amlogic's GPL kernel
  says that it's running at 1.28 MHz - it's actually programming a
  divider of 20 + 1. This is probably because their calculation is based
  on a 27 MHz XTAL clock, while the actual XTAL clock is 24 MHz)

None of these is a real showstopper as the ADC still works fine without
these changes. However, this is probably due to the fact that the
bootloader also does some initialization on the SAR ADC - and we don't
want to rely on that.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>