iio: adc: meson-saradc: add support for the Meson8m2 SoCs
The SAR ADC on Meson8m2 behaves identical to the one found in the
Meson8b SoCs. Support for the Meson8 and Meson8b is almost identical,
only the temperature sensor settings (trimming bits, multiplier and
divider) are different. Meson8m2 matches what we already had for
Meson8b, without this the voltage channels are working fine but the
temperature channel is reporting incorrect values.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>