Bluetooth: hci_h5: add support for the RTL8723BS bluetooth module - WiP
Realtek's RTL8723BS provides wifi and Bluetooth connectivity. The
Bluetooth module connects to the host using a serial console port.
This adds support for the Bluetooth module found in the RTL8723BS chip
by adding a serdev driver to hci_h5. The existing initialization
(firmware download and config patching) from "btrtl.c" is simply
re-used, as it seems to be compatible with the RTL8723BS chip.

WiP

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>