dt-bindings: usb: xhci: include the roothub and a device in the example
This extends the existing example from the USB xHCI binding
documentation so it includes the roothub and an actual device.
The goal of this is to show that the roothub is specified alongside the
actual devices on the USB bus (which is important because a device on
the USB bus - for example a hub - might need it's own phys / phy-names
properties. modelling the roothub as separate device and not nesting the
other devices on the bus below the roothub allows us to keep the
properties, for example the PHYs, separated).

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>