net: stmmac: dwmac-meson8b: document struct meson8b_dwmac
The IP block for the dwmac-meson8b glue supports RMII and RGMII modes.
Most of the struct meson8b_dwmac members should only be used in RGMII
mode though as the underlying register bits are RGMII mode specific.
Document the struct meson8b_dwmac members to make it clear in which mode
they should be used.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>