ARM: dts: meson8b: use the Meson8b Ethernet controller binding
Until now we have been using the "amlogic,meson6-dwmac" binding with the
register offset (0xc1108108) defined in meson.dtsi.
During testing (and by reading Hardkernel's u-boot sources for the
Odroid-C1) it turns out that the actual register that should be used is
at 0xc1108140. This also requires us to switch to the new
"amlogic,meson8b-dwmac" binding, because the dwmac-meson8b driver knows
how to configure that register. The old register is a no-op, so using
the old "amlogic,meson6-dwmac" binding with the old register meant that
we relied on the bootloader to set up the Ethernet clocks etc.
correctly.

Signed-off-by: Emiliano Ingrassia <ingrassia@epigenesys.com>
[add patch title/description]
Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>