ARM: dts: meson8b: Enable Ethernet on the Odroid-C1
This enables Ethernet on the Odroid-C1 which has an RTL8211F RGMII PHY.
The schematics show that the PHY's interrupt pin ("INTR") is connected
to GPIOH_3 and the reset line ("NRST") is routed to GPIOH_4

Testing shows that we need a TX delay of 4ns. With a delay of 2ns we can
receive data, but transmitted packets are lost (= they never arrive on
the other end, confirmed by using wireshark on the remote computer).

Signed-off-by: Emiliano Ingrassia <ingrassia@epigenesys.com>
[added patch title and description]
Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>